changed to zalando swagger-codegen plugin / splitted rest from business logic
